chore: move LiquidStakingToken
