+++ apps/tangle-dapp/components/LiquidStaking/stakeAndUnstake/ParachainWalletBalance.tsx
@@ -13,9 +13,9 @@ import { FC, useCallback, useMemo, useState } from 'react';
 import { twMerge } from 'tailwind-merge';
 
 import { EMPTY_VALUE_PLACEHOLDER } from '../../../constants';
-import { LiquidStakingToken } from '../../../constants/liquidStaking';
 import useParachainBalances from '../../../data/liquidStaking/useParachainBalances';
 import useSubstrateAddress from '../../../hooks/useSubstrateAddress';
+import { LiquidStakingToken } from '../../../types/liquidStaking';
 import formatBn from '../../../utils/formatBn';
 
 export type ParachainWalletBalanceProps = {
